Revert "feat: remove private prompts on related brain delete (#842)"
This reverts commit 4c15fe2bfde7a2fdc59c299ef668f1ba0cd8ffa8.
