--- frontend/app/brains-management/[brainId]/components/BrainManagementTabs/components/SettingsTab/components/PublicPrompts/components/PublicPromptsList/index.ts
@@ -1,1 +1,0 @@
-export * from "./PublicPromptsList";
