--- frontend/app/brains-management/[brainId]/components/BrainManagementTabs/components/SettingsTab/components/PublicPrompts/index.ts
@@ -1,1 +1,0 @@
-export * from "./PublicPrompts";
